Added simple input and changed spaceship color
I coded the main game container to react to input
from the keys that will be used in this game:
up, down, left, right, spacebar, and wasd.
Right now, the game simply reports key up and down
events through a basic text box at the top of the
container. I also changed the spaceship color to
something with more contrast to be easier on the eye