Added simple movement to asteroid object
I added simple movement to the asteroid object by
adding a fixed updater in the init code block that
updates the object's position each frame according
to a separate velocity property for both x and y.
I provided logic to make sure that the asteroid
switches to the other side of the game area when
reaching the border. I also moved the code for the
top bar lower in the main function to make sure it
stays on the top "layer" of objects on screen.